Cleanup unreachable UAnnotation branches
Previously, https://r.android.com/3577270 detected some untakeable
branches that assumed that a `UAnnotation` was a `UExpression`, or at
least that there was a class that implemented both. It did not appear
as though there was any such class. However, if there was such a class
then deleting the branch would silently ignore it. So, instead it made
each branch throw an exception to indicate that the branch was taken.

None of those exceptions have been thrown in that time showing those
branches have never been taken so this change removes them altogether.

Bug: 354633349
Test: ./gradlew
Change-Id: I6fadf3fb71677617f628b1ba84c550b63552f8e4