Update lint, AGP, gradle for metalava
The lint update pulls in a fix for a constant evaluation case that previously wasn't handled correctly, enabling removing some previously failing tests from the baseline for psi and changing the expected value on one test to the evaluated value (which was already the turbine result, so a turbine test was also removed from the baseline).

Updating AGP required also updating gradle.

Bug: 409203583
Test: ./gradlew
Change-Id: Ib9558a3f8514890d806475e581525f8040e96538