Disable checkSystemPermissions(MethodItem) in Android
The `checkSystemPermissions(MethodItem)` does not work in Android and
has not for at least 5 years but has been tested, to some extent, in
`SystemServiceCheckTest`. That behavior was caused because the code
attempted to match `@SystemApi` (without defaults) against
`@SystemApi(client = PRIVILEGED_APPS)` which failed and so
`checkSystemPermissions(MethodItem)` is never called on Android but
was called by `SystemServiceCheckTest`.

Change https://r.android.com/3599139 attempted to preserve that
previous behavior by using a `DefaultAnnotationItem` whose `toSource()`
method did not support `showDefaultAttrs = true` as that is what
added the defaults to `@SystemApi` so it would match against
`@SystemApi(client = PRIVILEGED_APPS)`. That ensured it would not call
`checkSystemPermissions(MethodItem)` on Android but would call it on
`SystemServiceCheckTest`.

However, change https://r.android.com/3602054 stops using `toSource()`
and makes `AnnotationFilterEntry` add the default values itself. That
means the code in https://r.android.com/3599139 will now match on
Android, so it will call `checkSystemPermissions(MethodItem)` which
will break the build by reporting errors, some of which are correct
but most of which are not.

This change stops using `@SystemApi` and adds a fake annotation that
is only matched by `SystemServiceCheckTest`. That ensures that change
https://r.android.com/3602054 will not break Android.

Bug: 412743564
Test: ./gradlew
Change-Id: Ic4a9bec9197ea430bd865a2428a6c54c73c4e1a4