Avoid resolving annotation class in Psi
In order to compute a `Value` for an annotation attribute it is
necessary to obtain the `TypeItem` for its attribute in order to
perform any necessary conversions.

Previously, Psi would use `CachingAnnotationValueProvider` to do that
by resolving the `AnnotationItem` to a `ClassItem` and searching for
the attribute's method and using its return type. That causes problems
if the `Value` is accessed during construction as it may end up
resolving a class which has not been created yet.

This change replaces the use of `CachingAnnotationValueProvider` with
a Psi specific class which does the equivalent job.

Bug: 354633349
Test: ./gradlew
Change-Id: I84ed4816b88258ac226cc24128858eafc9d50ae2