Stop checking if SdkConstantType is an EnumConstantValue
Previously, `SdkFileWriter` would check to make sure that the
`SdkConstant` annotation's `value` property was an `EnumConstantValue`.
While that is strictly correct it is not actually necessary as all that
matters is its `toValueString()` matches one of the hard-coded
constants. This change removes the check in preparation for merging
`EnumConstantValue` into `FieldReferenceValue`.

Bug: 413311223
Test: ./gradlew
Change-Id: Ie830ec91ea809dea5264ba0b52de72115db5dcd0