Add ATTRIBUTE_DEFAULT_FORMATTER and use in Turbine
Previously, Turbine had its own special method for creating the String
used for `legacyDefaultValue()`. This change removes that and replaces
it with a new `LegacyValueFormatter` instance called
`ATTRIBUTE_DEFAULT_FORMATTER`. That is hooked into
`DefaultMethodItem.legacyDefaultValue()` and will be called only if
`annotationDefault` is not provided (it defaults to `null`). That
allows Turbine to use the formatter by passing `null` while text
continues to supply its own String representation.

That fixed the `CommonParameterizedAttributeDefaultValueTest`
`testLegacySource()` tests that were previously failing on Turbine.

Bug: 354633349
Test: ./gradlew
Change-Id: I96e2305bc0fbff4ed53113af8c3f41fff9f09b6c