Support using source expression for ClassObjectValue
Psi does not specifically handle Kotlin class literal, it just falls
back to the source representation. In order to replicate that behavior
in the `LegacyValueFormatter` it is necessary to record the source
expression in the `ClassObjectValue` and use that if requested when
formatting it for Kotlin. This change adds support for that.

The changes were made by first adding `ValueKind.CLASS` to the list of
`ValueKind`s currently handled by `PsiMethodItem.legacyDefaultValue()`
and running the `CommonParameterizedAttributeDefaultValueTest` tests on
Psi to ensure that the class literal tests failed on Kotlin. Then the
changes were made to fix the tests.

Bug: 354633349
Test: ./gradlew
      # Tested in Android and AndroidX and does not affect signature
      # files in either or stubs in the former.
Change-Id: Ie69555c73aef79f7f999bd40d9f31c2f532ab4ad