Add missing classes to prebuilts/sdk test data
Metalava assumes that APIs are incremental, i.e. once added nothing is
ever really removed. Even when it is marked with `@removed` tha APIs
still exist. The test data for `prebuilts/sdk` did not follow those
rules which meant that it was an unrealistic test that could break
unexpectedly. This change adds the missing APIs to fix that.

Bug: 409725916
Test: ./gradlew
Change-Id: I53d47b0af9876f8a2778048402a9cad79266e3fc