Use DefaultMethodItem.legacyDefaultValue() for Psi Java
Previously, `PsiMethodItem` would provide its own implementation of
`legacyDefaultValue()` for Java and Kotlin. However, the
`DefaultMethodItem.legacyDefaultValue()` has equivalent behavior for
Java so this change removes the Psi code for Java.

This required a change to a test that used a cast in the default
expression. Previously, it would output the cast to the signature file
and stubs but after this change it just outputs the value. So,
`(double)Float.NEGATIVE_INFINITY` becomes `Double.NEGATIVE_INFINITY`.
The reason that happens is because the cast expression is converted to
its numerical value of `(-1.0/0.0)` and then on output it is mapped to
the field `Double.NEGATIVE_INFINITY`. This does not actually affect
either Android or AndroidX as neither uses a cast in this way.

Bug: 354633349
Test: ./gradlew
      # Tested in Android and AndroidX and does not affect signature
      # files in either or stubs in the former.
Change-Id: I76630a674148ec507dc15e2e86815a73c7d0f4ce