Fix couple of issues with KotlinTestClassCreator
There was an unnecessary `;` which the IDE reported as an error but
more importantly the `OtherAnnotation` definition used `emptyArray()`.
While that is allowed in Kotlin it is not recommended (the IDE flagged
that it should be replaced with a an array literal). The latter causes
problems for the following change.

The `PsiValueFactory` does not currently support that and it is not
clear whether it is necessary as AndroidX does not seem to use it. So,
rather than add support for it now, this change replaces it with an
array literal.

Bug: 354633349
Test: ./gradlew
      # Tested in Android and AndroidX and does not affect signature
      # files in either or stubs in the former.
Change-Id: I4a481246eed00a85cec487b2126cb6fac2014f5f