Support formatting Kotlin style AnnotationValues
Kotlin treats an annotation value, i.e. a nested `AnnotationItem` as
if it is normal class constructor, so no leading `@` and parentheses
are required, even if the attributes are empty. This change adds
support for producing those. It does that by adding a `ValueLanguage`
property to `ValueStringConfiguration`. That is currently only used for
annotations but will be used in future for `ClassObjectValue` too.

Will be used by a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia7b8a4629310980a2d086e31ecab46fdf6fb0122