Preserve the order of attributes in Kotlin AnnotationValues
Previously, `uCallExpressionToAnnotationValue(...)` would iterate over
an annotation class's constructor parameters, find the associated
argument and add an annotation attribute. That meant the attributes
were in the order of the parameters which could be different to the
argument order. However, the order of arguments matters.

This change flips the code to iterate over the arguments and find the
associated parameter which preserves the order. This is needed by a
follow-up change which relies on the attribute order.

Bug: 354633349
Test: ./gradlew
Change-Id: I91576261bfab32b832bddbf5243cb94ecb5f353e