Use LegacyValueFormatter in DefaultAnnotationItem.toSource()
Previously, `DefaultAnnotationItem.toSource()` would use the string
representation of `legacyValue` which did not match the behavior of the
`PsiAnnotationItem` and `UAnnotationItem` implementations, which for
all their inconsistencies defined the "correct" way to format them at
the moment.

This change adds a new `ANNOTATION_SOURCE_FORMATTER` to
`LegacyValueFormatter` and uses that to implement
`DefaultAnnotationItem.toSource()`.

It significant improves the compatibility between Turbine and Text and
the legacy Psi behavior as shown by the number of previously basedlined
tests that were fixed. Follow-up changes will iteratively improve upon
this.

Bug: 354633349
Test: ./gradlew
Change-Id: Ie3bb93ca0a83737cd245c704b6767f07784c85f3