Parse bytecode only APIs
This uses the pseudo-modifiers `@KotlinOnly`, `@BytecodeOnly`, `@InaccessibleFromKotlin`, and `@InaccessibleFromJava` to represent the different possible target language sets (besides the set of all languages, which uses no modifier). These are not processed as annotations, but using the `@` means that they will get syntax highlighting in gerrit. The plan for the modifiers was raised in the API council leads chat and approved.

Bug: 407735058
Test: ApiFileTest
Change-Id: Ie9e3a176e3ffeb5d3bb0ceb64b9742cda286cab4