Work around a missing java.lang.String class
The Android build will sometimes run Metalava without providing a
definition of `java.lang.String`, e.g. because it does not provide it
in the `--classpath` list or a `--jdk-home` option. That causes the
`PsiValueFactory` to fail to convert a string literal like `""` to a
`StringValue`. That is because the type is `String` not
`java.lang.String`.

That breakage is blocking over 30 other changes from being merged
across which is further blocking other work. This change works around
the issue by treating `String` as if it was `java.lang.String`.

Once the issue has been resolved this will be reverted.

Bug: 415333758
Test: ./gradlew
      # Run the following before and after merging this change.
      # Before merging it will fail after merging it will work.
      m out/soong/.intermediates/tools/tradefederation/core/tradefed-doc-stubs/linux_glibc_common/everything/tradefed-doc-stubs-stubs.srcjar
Change-Id: I2a0f72fa3f0dd1f29b74566c071dd008158ecca3