Handle bytecode-only APIs from JvmMultifileClasses
JvmMultifileClasses exist as several class files in a jar: one with the name of the multi-file class, which does not have Kotlin metadata about the class members, and one for each source file making up the class, which do have metadata about the class members. This change updates the jar processing to record the paths of the multi-file class parts in the first pass through all psi classes, and then take a second pass which loads the class parts.

Bug: 415369260
Test: CommonTargetLanguageTest
Change-Id: Ibdf45f6528ef27e1736476c0531ad940c4f4891d