Support parsing an array of annotations
Previously, when parsing an array `ValueParser` would retrieve a
general purpose token for the array element value. They use
parentheses as separators so it would split `{@Foo(attr=1)}` into:
  `{`, `@Foo`, `(`, `attr`, `=`, `1`, `}`
It would then treat `@Foo` as the first element and expect a `,` or `}`
token to follow but would find a `(` instead.

This change switches it to use a value token which does not use
parentheses as separators so it would split the above into:
`{`, `@Foo(attr=1)`, `}` which it could parse correctly.

Longer term the `parseArrayElementValue(...)` should be migrated to use
a `Tokenizer` so it does not have to parse the contents to extract a
token for the whole annotation and then parse that again and instead
could just parse it all in one go. However, this will do for now.

Bug: 354633349
Test: ./gradlew
Change-Id: Id3219308991d32271d1bee6f4c9a8422b024d7e0